requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
yfinance==0.2.40
numpy==1.26.4
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
APScheduler==3.10.4
bcrypt==4.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
